requirements.txt
fastapi
uvicorn
playwright
transformers
torch
python-dotenv
beautifulsoup4
pydantic
